+++ AWSChimeSDKMessaging/AWSChimeSDKMessagingService.m
@@ -25,7 +25,7 @@
 #import "AWSChimeSDKMessagingResources.h"
 
 static NSString *const AWSInfoChimeSDKMessaging = @"ChimeSDKMessaging";
-NSString *const AWSChimeSDKMessagingSDKVersion = @"2.25.0";
+NSString *const AWSChimeSDKMessagingSDKVersion = @"2.26.0";
 
 
 @interface AWSChimeSDKMessagingResponseSerializer : AWSJSONResponseSerializer