+++ AWSElasticLoadBalancing/AWSElasticLoadBalancingService.m
@@ -25,7 +25,7 @@
 #import "AWSElasticLoadBalancingResources.h"
 
 static NSString *const AWSInfoElasticLoadBalancing = @"ElasticLoadBalancing";
-NSString *const AWSElasticLoadBalancingSDKVersion = @"2.25.0";
+NSString *const AWSElasticLoadBalancingSDKVersion = @"2.26.0";
 
 
 @interface AWSElasticLoadBalancingResponseSerializer : AWSXMLResponseSerializer