+++ AWSSNS/AWSSNSService.m
@@ -25,7 +25,7 @@
 #import "AWSSNSResources.h"
 
 static NSString *const AWSInfoSNS = @"SNS";
-NSString *const AWSSNSSDKVersion = @"2.25.0";
+NSString *const AWSSNSSDKVersion = @"2.26.0";
 
 
 @interface AWSSNSResponseSerializer : AWSXMLResponseSerializer