+++ AWSSageMakerRuntime/AWSSageMakerRuntimeService.m
@@ -25,7 +25,7 @@
 #import "AWSSageMakerRuntimeResources.h"
 
 static NSString *const AWSInfoSageMakerRuntime = @"SageMakerRuntime";
-NSString *const AWSSageMakerRuntimeSDKVersion = @"2.25.0";
+NSString *const AWSSageMakerRuntimeSDKVersion = @"2.26.0";
 
 
 @interface AWSSageMakerRuntimeResponseSerializer : AWSJSONResponseSerializer